Result moved from 'task' to 'statistic' package
