+++ src/main/java/egl/client/controller/task/TaskController.java
@@ -3,7 +3,7 @@ package egl.client.controller.task;
 import java.util.function.Consumer;
 
 import egl.client.controller.Controller;
-import egl.core.model.task.Result;
+import egl.core.model.statistic.Result;
 import egl.core.model.task.Task;
 import egl.core.model.topic.Topic;
