+++ src/main/java/egl/client/controller/task/category/TestControllerTask4.java
@@ -4,7 +4,7 @@
 //import egl.client.model.topic.category.Category;
 //import egl.client.model.topic.category.Translation;
 //import egl.client.model.topic.category.Word;
-//import egl.core.model.task.Result;
+//import egl.core.model.statistic.Result;
 //import egl.core.model.task.Task;
 //import egl.core.model.topic.Topic;
 //import javafx.application.Platform;
